practice_007b_opentelemetry_dashboards/app/pyproject.toml
[project]
name = "otel-dashboards"
version = "0.1.0"
description = "OpenTelemetry metrics with Prometheus + Grafana dashboards"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-prometheus>=0.51b0",
    "prometheus-client>=0.22.0",
]
